+++ src/components/index.ts
@@ -1,9 +1,10 @@
+export { ActionItem } from './ActionItem';
 export { Block } from './Block';
 export { Fallback } from './Fallback';
 export { Layout } from './Layout';
 export { MenuDialog } from './Menu';
 export { Time } from './Time';
 export { UnitSwitch } from './UnitSwitch';
-export { ActionItem } from './ActionItem';
+export { Temperature } from './Temperature';
 
 export * from './Icons';
