+++ src/main.tsx
@@ -11,7 +11,7 @@ import App from './App.tsx';
 import { Layout } from './components';
 import { ErrorPage } from './components/ErrorPage.tsx';
 
-import { WelcomePage, SettingsPage, SearchPage, PlacesPage } from './components/Menu/Pages';
+import { WelcomePage, PlacesPage, SettingsPage, SearchPage } from './components/Menu/Pages';
 
 import { MenuPageRoutes } from './components/Menu/routes.ts';
